(automake): Revert to known good configure and Makefile.in
Previous rebase of branch lost some of chincheta's changes, which
broke the build for FreeBSD. Subsequent attempts to capture the
output of autoreconf on FreeBSD broke the build for everything else.
Revert ar-lib, test-driver, configure and src/Makefile.in to previous
versions within the 0.4 tarball.
Also, add Makefile.in, libtool and .dirstamps to .gitignore